requirements.txt
python-pptx>=0.6.23
lxml>=4.9.0
Pillow>=10.0.0
colormath>=3.0.0
numpy>=1.24.0
